Rename package from conf to config
