+++ addons/common/events/decisions/src/main/java/org/kie/kogito/eventdriven/decision/EventDrivenDecisionController.java
@@ -24,7 +24,7 @@ import java.util.concurrent.CompletionStage;
 import org.apache.commons.lang3.StringUtils;
 import org.kie.dmn.api.core.DMNContext;
 import org.kie.dmn.api.core.DMNResult;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.decision.DecisionExecutionIdUtils;
 import org.kie.kogito.decision.DecisionModel;
 import org.kie.kogito.decision.DecisionModels;
