+++ addons/common/events/predictions/src/main/java/org/kie/kogito/eventdriven/predictions/EventDrivenPredictionsController.java
@@ -23,7 +23,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 
 import org.kie.api.pmml.PMML4Result;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.event.DataEvent;
 import org.kie.kogito.event.DataEventFactory;
 import org.kie.kogito.event.EventEmitter;
