+++ addons/common/tracing/decision-common/src/main/java/org/kie/kogito/tracing/decision/aggregator/Aggregator.java
@@ -19,7 +19,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.kie.dmn.api.core.DMNModel;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.tracing.decision.event.evaluate.EvaluateEvent;
 
 import io.cloudevents.CloudEvent;
