+++ addons/common/tracing/decision-common/src/test/java/org/kie/kogito/tracing/decision/DecisionTracingCollectorTest.java
@@ -26,8 +26,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.kie.dmn.api.core.DMNModel;
 import org.kie.dmn.feel.util.Pair;
-import org.kie.kogito.conf.ConfigBean;
-import org.kie.kogito.conf.StaticConfigBean;
+import org.kie.kogito.config.ConfigBean;
+import org.kie.kogito.config.StaticConfigBean;
 import org.kie.kogito.event.cloudevents.utils.CloudEventUtils;
 import org.kie.kogito.tracing.decision.event.evaluate.EvaluateEvent;
 import org.kie.kogito.tracing.decision.mock.MockDefaultAggregator;
