+++ kogito-codegen-modules/kogito-codegen-core/src/main/resources/class-templates/config/ConfigBeanJavaTemplate.java
@@ -15,6 +15,6 @@
  */
 package $Package$;
 
-public class ConfigBean extends org.kie.kogito.conf.StaticConfigBean {
+public class ConfigBean extends org.kie.kogito.config.StaticConfigBean {
 
 }
