+++ kogito-codegen-modules/kogito-codegen-core/src/main/resources/class-templates/config/ConfigBeanSpringTemplate.java
@@ -16,14 +16,14 @@
 package $Package$;
 
 @org.springframework.stereotype.Component
-public class ConfigBean extends org.kie.kogito.conf.StaticConfigBean {
+public class ConfigBean extends org.kie.kogito.config.StaticConfigBean {
 
     @org.springframework.beans.factory.annotation.Value("${kogito.service.url:#{null}}")
     java.util.Optional<java.lang.String> kogitoService;
 
     @org.springframework.beans.factory.annotation.Value("${kogito.messaging.as-cloudevents:#{true}}")
     boolean useCloudEvents;
-    
+
     @org.springframework.beans.factory.annotation.Value("${kogito.jackson.fail-on-empty-bean:#{false}}")
     boolean failOnEmptyBean;
