+++ kogito-codegen-modules/kogito-codegen-core/src/main/resources/class-templates/config/GlobalObjectMapperQuarkusTemplate.java
@@ -33,7 +33,7 @@ import io.quarkus.jackson.ObjectMapperCustomizer;
 public class GlobalObjectMapper implements ObjectMapperCustomizer {
     
     @Inject
-    org.kie.kogito.conf.ConfigBean configBean;
+    org.kie.kogito.config.ConfigBean configBean;
 
     public void customize(ObjectMapper mapper) {
         if (!configBean.failOnEmptyBean()) {
