+++ quarkus/addons/events/rules/runtime/src/main/java/org/kie/kogito/eventdriven/rules/QuarkusEventDrivenRulesController.java
@@ -18,7 +18,7 @@ package org.kie.kogito.eventdriven.rules;
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.event.EventEmitter;
 import org.kie.kogito.event.EventReceiver;
