+++ quarkus/addons/monitoring/core/src/main/java/org/kie/kogito/monitoring/core/quarkus/QuarkusEventListenerFactory.java
@@ -20,7 +20,7 @@ import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 
 import org.kie.kogito.KogitoGAV;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.drools.core.config.DefaultRuleEventListenerConfig;
 import org.kie.kogito.internal.process.event.KogitoProcessEventListener;
 import org.kie.kogito.monitoring.core.common.Constants;
