+++ quarkus/extensions/kogito-quarkus-extension-common/kogito-quarkus-common/src/main/java/org/kie/kogito/quarkus/config/KogitoEventingRuntimeConfig.java
@@ -14,38 +14,28 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.quarkus.conf;
-
-import java.util.Optional;
+package org.kie.kogito.quarkus.config;
 
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
 
 @ConfigGroup
-public class KogitoPersistenceRuntimeConfig {
+public class KogitoEventingRuntimeConfig {
+
+    private static final String MAX_THREADS_PROPERTY = "threads.poolSize";
+    private static final String DEFAULT_MAX_THREADS = "10";
+    private static final String DEFAULT_QUEUE_SIZE = "1";
+    private static final String QUEUE_SIZE_PROPERTY = "threads.queueSize";
 
     /**
-     * Persistence DB type
+     * Maximum number of threads to handle incoming events by channel
      */
-    @ConfigItem
-    public Optional<PersistenceType> type;
+    @ConfigItem(name = MAX_THREADS_PROPERTY, defaultValue = DEFAULT_MAX_THREADS)
+    public int maxThreads;
 
     /**
-     * Automatically apply database schema changes
+     * Maximum size of queue to hold incoming events to be processed by threads in the pool
      */
-    @ConfigItem(name = "auto.ddl", defaultValue = "true")
-    public boolean autoDDL;
-
-    /**
-     * Use optimistic locking
-     */
-    @ConfigItem(name = "optimistic.lock", defaultValue = "false")
-    public boolean optimisticLock;
-
-    /**
-     * Query execution timeout
-     */
-    @ConfigItem(name = "query.timeout.millis", defaultValue = "10000")
-    public long queryTimeout;
-
+    @ConfigItem(name = QUEUE_SIZE_PROPERTY, defaultValue = DEFAULT_QUEUE_SIZE)
+    public int queueSize;
 }
