+++ quarkus/extensions/kogito-quarkus-extension-common/kogito-quarkus-common/src/main/java/org/kie/kogito/quarkus/config/KogitoPersistenceBuildTimeConfig.java
@@ -14,28 +14,24 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.quarkus.conf;
+package org.kie.kogito.quarkus.config;
 
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
 
 @ConfigGroup
-public class KogitoEventingRuntimeConfig {
-
-    private static final String MAX_THREADS_PROPERTY = "threads.poolSize";
-    private static final String DEFAULT_MAX_THREADS = "10";
-    private static final String DEFAULT_QUEUE_SIZE = "1";
-    private static final String QUEUE_SIZE_PROPERTY = "threads.queueSize";
+public class KogitoPersistenceBuildTimeConfig {
 
     /**
-     * Maximum number of threads to handle incoming events by channel
+     * Generate Protobuf marshallers for runtime
      */
-    @ConfigItem(name = MAX_THREADS_PROPERTY, defaultValue = DEFAULT_MAX_THREADS)
-    public int maxThreads;
+    @ConfigItem(name = "proto.marshaller", defaultValue = "true")
+    public boolean runtimeProtoMarshaller;
 
     /**
-     * Maximum size of queue to hold incoming events to be processed by threads in the pool
+     * Generate Protobuf marshallers for Data Index
      */
-    @ConfigItem(name = QUEUE_SIZE_PROPERTY, defaultValue = DEFAULT_QUEUE_SIZE)
-    public int queueSize;
+    @ConfigItem(name = "data-index.proto.generation", defaultValue = "true")
+    public boolean dataIndexProtoMarshaller;
+
 }
