+++ quarkus/extensions/kogito-quarkus-extension-common/kogito-quarkus-common/src/main/java/org/kie/kogito/quarkus/config/PersistenceType.java
@@ -13,25 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.quarkus.conf;
 
-import io.quarkus.runtime.annotations.ConfigGroup;
-import io.quarkus.runtime.annotations.ConfigItem;
+package org.kie.kogito.quarkus.config;
 
-@ConfigGroup
-public class KogitoRPCRuntimeConfig {
-
-    /**
-     * Indicates if default value of enumerations should be included in grpc response
-     * 
-     */
-    @ConfigItem(name = "enum.includeDefault", defaultValue = "false")
-    public boolean enumDefault;
-
-    /**
-     * Time to wait for response from server when using streams
-     * 
-     */
-    @ConfigItem(name = "stream.timeout", defaultValue = "20")
-    public int streamTimeout;
+public enum PersistenceType {
+    JDBC,
+    MONGODB,
+    FILESYSTEM,
+    KAFKA,
+    INFINISPAN,
+    POSTGRESQL
 }
