+++ quarkus/extensions/kogito-quarkus-processes-extension/kogito-quarkus-processes/src/main/java/org/kie/kogito/quarkus/processes/KogitoBeanProducer.java
@@ -17,7 +17,7 @@ package org.kie.kogito.quarkus.processes;
 
 import javax.enterprise.inject.Produces;
 
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.correlation.CorrelationService;
 import org.kie.kogito.event.correlation.DefaultCorrelationService;
 import org.kie.kogito.process.ProcessVersionResolver;
