+++ springboot/addons/events/decisions/src/main/java/org/kie/kogito/eventdriven/decision/SpringBootEventDrivenDecisionController.java
@@ -17,7 +17,7 @@ package org.kie.kogito.eventdriven.decision;
 
 import javax.annotation.PostConstruct;
 
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.decision.DecisionModels;
 import org.kie.kogito.event.EventEmitter;
 import org.kie.kogito.event.EventReceiver;
