+++ springboot/addons/events/predictions/src/main/java/org/kie/kogito/eventdriven/predictions/SpringBootEventDrivenPredictionsController.java
@@ -17,7 +17,7 @@ package org.kie.kogito.eventdriven.predictions;
 
 import javax.annotation.PostConstruct;
 
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.event.EventEmitter;
 import org.kie.kogito.event.EventReceiver;
 import org.kie.kogito.prediction.PredictionModels;
