+++ springboot/addons/monitoring/core/src/main/java/org/kie/kogito/monitoring/core/springboot/SpringbootSystemMetricsCollectorProvider.java
@@ -18,7 +18,7 @@ package org.kie.kogito.monitoring.core.springboot;
 import javax.annotation.PostConstruct;
 
 import org.kie.kogito.KogitoGAV;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 import org.kie.kogito.monitoring.core.common.system.metrics.SystemMetricsCollector;
 import org.kie.kogito.monitoring.core.common.system.metrics.SystemMetricsCollectorProvider;
 import org.springframework.beans.factory.annotation.Autowired;
