+++ springboot/addons/tracing-decision/src/test/java/org/kie/kogito/tracing/decision/BaseSpringBootDecisionTracingTest.java
@@ -23,8 +23,8 @@ import org.junit.jupiter.api.Test;
 import org.kie.dmn.api.core.DMNContext;
 import org.kie.dmn.api.core.DMNRuntime;
 import org.kie.kogito.Application;
-import org.kie.kogito.conf.ConfigBean;
-import org.kie.kogito.conf.StaticConfigBean;
+import org.kie.kogito.config.ConfigBean;
+import org.kie.kogito.config.StaticConfigBean;
 import org.kie.kogito.decision.DecisionModel;
 import org.kie.kogito.decision.DecisionModels;
 import org.kie.kogito.dmn.DMNKogito;
