+++ springboot/addons/tracing-decision/src/test/java/org/kie/kogito/tracing/decision/SpringBootDecisionTracingConfigurationTest.java
@@ -17,7 +17,7 @@ package org.kie.kogito.tracing.decision;
 
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.Application;
-import org.kie.kogito.conf.ConfigBean;
+import org.kie.kogito.config.ConfigBean;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
